+++ unittests/Core/BuildEngineTest.cpp
@@ -87,7 +87,7 @@ public:
 
     // Request all of the inputs.
     inputValues.resize(inputs.size());
-    for (int i = 0, e = inputs.size(); i != e; ++i) {
+    for (unsigned long i = 0, e = inputs.size(); i != e; ++i) {
       engine.taskNeedsInput(this, inputs[i], i);
     }
   }
